requirements.txt
langchain
langgraph
langchain-groq
sentence-transformers
python-dotenv
numpy